+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ### Features
 1. [#49](https://github.com/influxdata/influxdb-client-php/pull/49): Added support for Guzzle 7
 1. [#51](https://github.com/influxdata/influxdb-client-php/pull/51): CSV parser is able to parse export from UI
+1. [#53](https://github.com/influxdata/influxdb-client-php/pull/53): Ability to write to DB via UDP protocol
 
 ### Documentation